Moved all paths to test files into a single "fixtures" module
